+++ src/main/java/com/example/food_app/dto/request/user/OrderItemRequest.java
@@ -5,7 +5,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @Data
@@ -13,8 +12,8 @@ import java.util.List;
 @NoArgsConstructor
 @Builder
 public class OrderItemRequest {
-    private BigInteger menuId;
-    private BigInteger menuSizeId;
-    private List<BigInteger> toppingIds;
+    private Long menuId;
+    private Long menuSizeId;
+    private List<Long> toppingIds;
     private Integer quantity;
 }
