+++ src/main/java/com/example/food_app/dto/request/user/OrderRequest.java
@@ -3,7 +3,6 @@ package com.example.food_app.dto.request.user;
 import com.example.food_app.entity.enums.PaymentMethod;
 import lombok.*;
 
-import java.math.BigInteger;
 
 @Data
 @AllArgsConstructor
@@ -13,7 +12,7 @@ public class OrderRequest {
     private PaymentMethod paymentMethod;
     private String fullName;
     private String phone;
-    private BigInteger wardId;
+    private Long wardId;
     private String addressDetail;
     private String note;
     private String voucherCode;
