+++ src/main/java/com/example/food_app/dto/request/user/ReviewItemRequest.java
@@ -5,14 +5,13 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigInteger;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
 public class ReviewItemRequest {
-    private BigInteger orderDetailId;
+    private Long orderDetailId;
     private Float rating;
     private String comment;
 }
