+++ src/main/java/com/example/food_app/dto/response/admin/AccountResponse.java
@@ -6,7 +6,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 @Data
@@ -14,7 +13,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @Builder
 public class AccountResponse {
-    private BigInteger accountId;
+    private Long accountId;
     private String fullName;
     private String email;
     private String phone;
