+++ src/main/java/com/example/food_app/dto/response/admin/CategoryResponse.java
@@ -5,7 +5,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 @Data
@@ -13,7 +12,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @Builder
 public class CategoryResponse {
-    private BigInteger categoryId;
+    private Long categoryId;
     private String name;
     private String description;
     private String iconUrl;
