+++ src/main/java/com/example/food_app/dto/response/admin/WardResponse.java
@@ -5,7 +5,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 @Data
@@ -13,7 +12,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @Builder
 public class WardResponse {
-    private BigInteger wardId;
+    private Long wardId;
     private String wardCode;
     private String name;
     private Integer shippingFee;
