+++ src/main/java/com/example/food_app/dto/response/user/ProfileResponse.java
@@ -5,7 +5,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 @Data
@@ -13,7 +12,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @Builder
 public class ProfileResponse {
-    private BigInteger accountId;
+    private Long accountId;
     private String fullName;
     private String email;
     private String phone;
