+++ src/main/java/com/example/food_app/dto/response/user/Voucher1Response.java
@@ -6,7 +6,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 @Data
@@ -14,7 +13,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @Builder
 public class Voucher1Response {
-    private BigInteger voucherId;
+    private Long voucherId;
     private String code;
     private String title;
     private String description;
