+++ src/main/java/com/example/food_app/dto/response/user/VoucherResponse.java
@@ -7,7 +7,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 @Data
@@ -15,7 +14,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @Builder
 public class VoucherResponse {
-    private BigInteger voucherId;
+    private Long voucherId;
     private String code;
     private String title;
     private String description;
