+++ src/main/java/com/example/food_app/entity/Cart.java
@@ -8,7 +8,6 @@ import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -21,7 +20,7 @@ import java.util.List;
 public class Cart {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private BigInteger cartId;
+    private Long cartId;
 
     @OneToOne
     @JoinColumn(name = "account_id", nullable = false, unique = true)
