+++ src/main/java/com/example/food_app/entity/FlashSale.java
@@ -7,7 +7,6 @@ import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -21,7 +20,7 @@ public class FlashSale {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "flash_sale_id")
-    private BigInteger flashSaleId;
+    private Long flashSaleId;
 
     private String title;
