+++ src/main/java/com/example/food_app/entity/Menu.java
@@ -8,7 +8,6 @@ import org.hibernate.annotations.UpdateTimestamp;
 import org.hibernate.type.SqlTypes;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.*;
 
@@ -22,7 +21,7 @@ public class Menu {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "menu_id")
-    private BigInteger menuId;
+    private Long menuId;
 
     @ManyToOne
     @JoinColumn(name = "category_id")
