+++ src/main/java/com/example/food_app/entity/OrderDetail.java
@@ -7,7 +7,6 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -21,7 +20,7 @@ public class OrderDetail {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "order_detail_id")
-    private BigInteger orderDetailId;
+    private Long orderDetailId;
 
     @ManyToOne
     @JoinColumn(name = "order_id")
