+++ src/main/java/com/example/food_app/repository/CartRepository.java
@@ -4,9 +4,8 @@ import com.example.food_app.entity.Account;
 import com.example.food_app.entity.Cart;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
-public interface CartRepository extends JpaRepository<Cart, BigInteger> {
+public interface CartRepository extends JpaRepository<Cart, Long> {
     Optional<Cart> findByAccount(Account account);
 }
