+++ src/main/java/com/example/food_app/repository/FlashSaleRepository.java
@@ -4,9 +4,8 @@ import com.example.food_app.entity.FlashSale;
 import com.example.food_app.entity.Voucher;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
 import java.util.List;
 
-public interface FlashSaleRepository extends JpaRepository<FlashSale, BigInteger> {
+public interface FlashSaleRepository extends JpaRepository<FlashSale, Long> {
     List<FlashSale> findAllByOrderByCreatedAtDesc();
 }
